Added a way to display registered data parsing tasks to the frontend. Planning to have orchestrator dynamically select data processing task on a datacell by datacell basis based on the task name registered in our column definition. This will make it far more practical for devs to create custom extractors.
