+++ config/graphql/queries.py
@@ -1,10 +1,13 @@
+import inspect
 import logging
 import re
 
 import graphene
+from celery.exceptions import CeleryError
 from django.conf import settings
 from django.db.models import Q
 from graphene import relay
+from graphene.types.generic import GenericScalar
 from graphene_django.fields import DjangoConnectionField
 from graphene_django.filter import DjangoFilterConnectionField
 from graphql_jwt.decorators import login_required
@@ -811,3 +814,43 @@ class Query(graphene.ObjectType):
             return Datacell.objects.filter(
                 Q(extract__creator=info.context.user) | Q(is_public=True)
             )
+
+    registered_extract_tasks = graphene.Field(
+        GenericScalar
+    )
+
+    @login_required
+    def resolve_registered_extract_tasks(self, info, **kwargs):
+        from config import celery_app
+
+        tasks = {}
+
+        # Try to get tasks from the app instance
+        # Get tasks from the app instance
+        try:
+            for task_name, task in celery_app.tasks.items():
+                if not task_name.startswith('celery.'):
+                    docstring = inspect.getdoc(task.run) or "No docstring available"
+                    tasks[task_name] = docstring
+
+        except AttributeError as e:
+            logger.warning(f"Couldn't get tasks from app instance: {str(e)}")
+
+        # Saving for reference... but I don't think it's necessary ATM and it's much higher latency.
+        # Try to get tasks from workers
+        # try:
+        #     i = celery_app.control.inspect(timeout=5.0, connect_timeout=5.0)
+        #     registered_tasks = i.registered()
+        #     if registered_tasks:
+        #         for worker_tasks in registered_tasks.values():
+        #             for task_name in worker_tasks:
+        #                 if not task_name.startswith('celery.') and task_name not in tasks:
+        #                     # For tasks only found on workers, we can't easily get the docstring
+        #                     tasks[task_name] = "Docstring not available for worker-only task"
+        # except CeleryError as e:
+        #     logger.warning(f"Celery error while inspecting workers: {str(e)}")
+        # except Exception as e:
+        #     logger.warning(f"Unexpected error while inspecting workers: {str(e)}")
+
+        # Filter out Celery's internal tasks
+        return {task: description for task, description in tasks.items() if task.startswith('opencontractserver.tasks.data_extract_tasks')}
