+++ opencontractserver/tasks/__init__.py
@@ -1,7 +1,8 @@
 from .cleanup_tasks import delete_analysis_and_annotations_task
 from .doc_tasks import burn_doc_annotations
 from .export_tasks import package_annotated_docs
-from .extract_tasks import llama_index_doc_query, run_extract
+from .extract_orchestrator_tasks import run_extract
+from .data_extract_tasks import *
 from .fork_tasks import fork_corpus
 from .import_tasks import import_corpus, import_document_to_corpus
 from .lookup_tasks import build_label_lookups_task
@@ -17,7 +18,6 @@ from .query_tasks import run_query
 __all__ = [
     "run_query",
     "run_extract",
-    "llama_index_doc_query",
     "package_annotated_docs",
     "burn_doc_annotations",
     "fork_corpus",
